Move `PodmanArgs` to its own module
